+++ digiarch/commands/edit/edit.py
@@ -2,6 +2,7 @@ from click import group
 
 from .action import group_action
 from .lock import command_lock
+from .processed import command_processed
 from .remove import command_remove
 from .rename import command_rename
 from .rollback import command_rollback
@@ -48,6 +49,7 @@ def group_edit():
 group_edit.add_command(group_action, group_action.name)
 group_edit.add_command(command_rename, command_rename.name)
 group_edit.add_command(command_lock, command_lock.name)
+group_edit.add_command(command_processed, command_processed.name)
 group_edit.add_command(command_remove, command_remove.name)
 group_edit.add_command(command_rollback, command_rollback.name)