+++ digiarch/commands/extract/extractors/extractor_zip.py
@@ -7,7 +7,7 @@ from zipfile import ZipFile
 
 from acacore.utils.functions import rm_tree
 
-from digiarch.doctor import sanitize_path
+from digiarch.commands.doctor import sanitize_path
 
 from .base import ExtractError
 from .base import ExtractorBase